move functions in utils.py into directory service, add first cache
